Create Product Category models.
